Documentation Addressing Integer Primary Key
This PR is a continuation of the discussion on the issue opened at `duckdb` about `SERIAL` datatypes. It might be helpful to include a snippet in the README to address the missing support for PostgreSQL-only datatypes like `SERIAL` and provide workarounds (like the `Sequence()` used in the existing example). I've written up some examples in this PR as suggestions.