fix: propagate kwargs through compilation nodes
Fixes https://github.com/Mause/duckdb_engine/issues/1138

When we visit a struct or union, we are handed an IdentifierProvider. Before, we just used it at this first layer, but forgot to pass it down to lower compilation steps. Now, we pass it along. This was a problem when we encountered nested dtypes, eg struct<struct<>>